Disable service worker
An nginx configuration update should now allow for direct access of
non-root pages to be delegated to React.

This does not fix redirects going to the wrong place (after refresh).

The rest of the service worker code is still required to ensure that
other devices (like my phone, and Ami's laptop) unregister the service
worker instead of continually deploying a very old version of the
website.